Addition to main page
Placed the progress bar on top of a no-show button. The addition is
nice, it allows the user more area to click on rather than a thin strip
where you can't tell whether or not you're clicking the bar. Minor
changes were made to the progress bar and percentage to fit the button
better. I tried to add in the transactions for the expense overview,
however, the code made Basil crash so it's commented out.